requirements.txt
Django==3.2.6
django-crontab==0.7.1
django-simple-history==3.1.1
Pillow==9.3.0
python-decouple==3.6
requests==2.28.1
xlrd==2.0.1